fix all the stupid things
