+++ src/extensions/extensions.ts
@@ -1,10 +1,38 @@
-import { Extension } from '../workers/extension-host'
 import { basename } from 'path'
 
-// TODO: move this to some more centralized place and share with other modules
-interface Disposable {
+export interface ExtensionInfo {
+  name: string
+  publisher: string
+}
+
+export enum ActivationEventType {
+  WorkspaceContains = 'workspaceContains',
+  Language = 'onLanguage',
+  Command = 'onCommand',
+  Debug = 'onDebug',
+  DebugInitialConfigs = 'onDebugInitialConfigurations',
+  DebugResolve = 'onDebugResolve',
+  View = 'onView',
+  Always  = '*',
+}
+
+interface ActivationEvent {
+  type: ActivationEventType
+  value: string
+}
+
+export interface Disposable {
   dispose: () => any
   [index: string]: any
+}
+
+export interface Extension extends ExtensionInfo {
+  config: any
+  packagePath: string
+  requirePath: string
+  extensionDependencies: string[]
+  activationEvents: ActivationEvent[]
+  subscriptions: Set<Disposable>
 }
 
 export const activateExtension = async (e: Extension): Promise<Disposable[]> => {
